Split chat view and state
